refactor: resize함수 useWindowResize 훅으로 재작성
